refactor[speculative]: simplify Eagle3 data module structure and reduce code duplication
